Moving configuration to its own file
